extracted ordersPane as a component for maintainability
